requirements.txt
fastapi>=0.111.0
uvicorn[standard]>=0.29.0
asyncpg>=0.29.0
pydantic>=2.7.0
python-multipart>=0.0.9
mcp>=1.0.0
anthropic>=0.40.0
pdfplumber>=0.11.0
python-dotenv>=1.0.0